requirements.txt
requests==2.9.1
slackbot==0.3.1
pyapi-gitlab==7.8.5
